Continue directory and database scaffolding

+++ DataManager.cs
@@ -1,5 +1,11 @@
+using System;
+using System.Net;
+using System.Reflection.Metadata;
+using System.Collections.Generic;
 using Microsoft.Data.Sqlite;
+
 namespace External;
+
 public class DataManager
 {
     internal void InitFilesystem()
@@ -8,12 +14,20 @@ public class DataManager
         try
         {
             // create user directory (dont need to check if it does or doesnt exist)
-            Directory.CreateDirectory("../../../User");
-            
+            Directory.CreateDirectory(GDirectories.userPath);
+            Directory.CreateDirectory(GDirectories.playerPath);
+
             // create user db
-            //using (FileStream fs = File.Create(userDBPath)) ;
-            FileStream fs = File.Create(GDirectories.userDBPath);
-            fs.Close();
+            if (!File.Exists(GDirectories.userDBPath))
+            {
+                File.Create(GDirectories.userDBPath);
+            }
+            else if (!File.Exists(GDirectories.playerDBPath))
+            {
+                File.Create(GDirectories.playerDBPath);
+            }
+            // using (var userFileStream = File.Create(GDirectories.userDBPath)) ;
+            // using (var playerFileStream = File.Create(GDirectories.playerPath)) ;
         }
         catch (Exception e)
         {
@@ -21,21 +35,8 @@ public class DataManager
         }
     }
 
-    internal void InititDB()
+    internal void VerifyUser()
     {
-        DBManager DBM = new DBManager();
-
-        try
-        {
-            DBM.ConnectToDB();
-        }
-        catch (Exception e)
-        {
-            Console.WriteLine(e);
-            throw;
-        }
-        
-        
     }
 }
 
@@ -51,22 +52,107 @@ public class DBManager
         //DBConnection.Close();
     }
 
-    internal void ConnectToDB()
+    internal void InitUserDB()
     {
-        using (var connection = new SqliteConnection(@"Data Source=../../../User/user.db"))
+        try
         {
-            Console.WriteLine("Yay!");
+            using (var conn = new SqliteConnection($@"Data Source={GDirectories.userDBPath}"))
+            {
+                conn.Open();
+                Debug.WDMNL("Connection Initialised: userDB");
+                var comm = conn.CreateCommand();
+                comm.CommandText =
+                    @"
+                    CREATE TABLE IF NOT EXISTS users (
+                        id INTEGER PRIMRARY KEY,
+                        firstname TEXT NOT NULL,
+                        lastname TEXT NOT NULL,
+                        age INTEGER NOT NULL 
+                    );
+                    ";
+                var readr = comm.ExecuteNonQuery();
+            }
+        }
+        catch (Exception e)
+        {
+            throw new DatabaseConnectionFailed(e);
         }
     }
 
-    internal void GetDBDate()
+    internal void InitPlayerDB()
     {
-        
+        try
+        {
+            using (var conn = new SqliteConnection($@"Data Source={GDirectories.userDBPath}"))
+            {
+                Debug.WDMNL("Connection Initialsed: playerDB");
+                var comm = conn.CreateCommand();
+                comm.CommandText =
+                    @"
+                    CREATE TABLE IF NOT EXISTS players (
+                        id INTEGER PRIMARY KEY,
+                        userid INTEGER NOT NULL,
+                        playername TEXT NOT NULL,
+                        playerclass TEXT NOT NULL,
+                    )
+                    ";
+            }
+        }
+        catch (Exception e)
+        {
+            throw new DatabaseConnectionFailed(e);
+        }
+    }
+
+    internal List<string> GetUserDBData()
+    {
+        List<string> userDBData = new List<string>();
+
+        using (var conn = new SqliteConnection($@"Data Source={GDirectories.userDBPath}"))
+        {
+            conn.Open();
+            var comm = conn.CreateCommand();
+            comm.CommandText =
+                @"
+                    SELECT * FROM users;
+                ";
+            using (var readr = comm.ExecuteReader())
+            {
+                while (readr.Read())
+                {
+                    userDBData.Add(readr.GetString(0));
+                }
+            }
+        }
+
+        return userDBData;
+    }
+
+    internal string GetPlayerDBData()
+    {
+        return "";
+    }
+
+    internal void WriteToUserDB(User user)
+    {
+        using (var conn = new SqliteConnection($@"Data Source={GDirectories.userDBPath}"))
+        {
+            conn.Open();
+            var comm = conn.CreateCommand();
+            comm.CommandText =
+                $@"
+                    INSERT INTO users (firstname, lastname, age) VALUES ({user.firstname}, {user.lastname}, {user.age});
+                ";
+            using (var readr = comm.ExecuteReader());
+
+        }
     }
 }
 
 public static class GDirectories
 {
     public static string userPath = @"../../../User/";
-    public static string userDBPath = @"../../../User/user.db";
+    public static string userDBPath = @"../../../User/user.sqlite";
+    public static string playerPath = @"../../../Player/";
+    public static string playerDBPath = @"../../../Player/player.sqlite";
 }
